+++ src/node/connection_types.cpp
@@ -2,6 +2,7 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
+#include <common/transport.h>
 #include <node/connection_types.h>
 #include <cassert>
